+++ packages/deepyr/lib/src/base/component_stylings.dart
@@ -92,6 +92,7 @@ mixin AllComponentStylings
         DropdownContentStyling,
         FieldsetStyling,
         FileInputStyling,
+        FilterStyling,
         FooterStyling,
         FormLabelStyling,
         HeroStyling,